[MOD] : SavedSpaceViewHolder 분리 #130
